BRABO-25: Processed upsource review

+++ src/main/java/org/brabocoin/brabocoin/node/config/BraboConfig.java
@@ -21,5 +21,5 @@ public interface BraboConfig {
      *
      * @return The maximum file size.
      */
-    long maxBlockFileSize();
+    int maxBlockFileSize();
 }
