+++ src/test/java/org/brabocoin/brabocoin/testutil/MockBraboConfig.java
@@ -42,7 +42,7 @@ public class MockBraboConfig implements BraboConfig {
     }
 
     @Override
-    public long maxBlockFileSize() {
+    public int maxBlockFileSize() {
         return delegator.maxBlockFileSize();
     }
 }
